requirements.txt
python-dotenv==1.0.0
cloudinary==1.36.0

